Declare AddedFriends array
This array will store all ids of friends, which are retrieved from the
chats object created in the App component. The useMemo hook is used
to prevent running that opeation every time the component is rendered.

The AddedFriends array is used to distinguish the users that are already
friends of the current user.